refactor(cli): extract eval/learn/obsidian handlers into scripts/cli/ modules
ARCH-1: main() was a ~1000-line function with command implementations inline (the eval
case alone ~561 lines), and cli.js exceeded the 700-line hard limit. Moved the three large
handlers into scripts/cli/ (CLI tier — not the scripts/lib engine tier): eval-command.js
(runEvalCommand + parseK/formatStatus/requireScenario closures), learn-command.js,
obsidian-command.js, plus shared.js (output). main() is now a thin dispatcher; switch
breaks became returns; try/catch preserved so handler throws still reach main()'s catch.

cli.js 1391 -> 657 lines (under the 700 hard limit). Source moved verbatim; full npm test
green (scripts 1831, hooks 213, node 14+7, skills 546, observer-daemon 17), lint clean.
